+++ NowChat-Rest/src/main/java/com/ghj/rest/model/Message.java
@@ -10,7 +10,8 @@ import java.util.Date;
  * @author 
  */
 public class Message implements Serializable {
-    private Integer id;
+
+    private Long id;
 
     /**
      * 接收状态
@@ -38,11 +39,11 @@ public class Message implements Serializable {
 
     private static final long serialVersionUID = 1L;
 
-    public Integer getId() {
+    public Long getId() {
         return id;
     }
 
-    public void setId(Integer id) {
+    public void setId(Long id) {
         this.id = id;
     }
